+++ packages/ui/lib/components/ui/index.ts
@@ -1,3 +1,4 @@
+export { Badge, badgeVariants, type BadgeProps } from './badge';
 export { Button, buttonVariants, type ButtonProps } from './button';
 export { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from './card';
 export {
@@ -26,5 +27,6 @@ export {
 } from './select';
 export { ScrollArea, ScrollBar } from './scroll-area';
 export { Switch } from './switch';
+export { TagsInput, type TagsInputProps } from './tags-input';
 export { Textarea } from './textarea';
 export { Toast, useToast, type ToastProps, type ToastType } from './toast';